pyproject.toml
[project]
name = "pmi-retail-workshop"
version = "0.1.0"
description = "PMI Retail Execution Workshop - Snowflake MDM & AI Agents"
authors = [
    { name = "PMI Workshop Team", email = "workshop@pmi.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
keywords = ["retail", "mdm", "ai", "snowflake", "langchain", "pmi"]

dependencies = [
    # Database connections
    "snowflake-connector-python>=3.7.0",
    "pandas>=2.1.0",
    "sqlalchemy>=2.0.0",
    # AI and LangChain
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.0.20",
    "openai>=1.10.0",
    # Web frameworks
    "streamlit>=1.29.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    # Data processing
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "plotly>=5.17.0",
    "seaborn>=0.12.0",
    "matplotlib>=3.7.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "loguru>=0.7.0",
    "httpx>=0.25.0",
    "faiss-cpu>=1.12.0",
    "pypdf>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
pmi-setup = "pmi_retail.scripts.setup:main"
pmi-demo = "pmi_retail.scripts.demo:main"
pmi-agents = "pmi_retail.agents.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pmi_retail"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short --cov=src/pmi_retail --cov-report=term-missing"
